pyproject.toml
[project]
name = "WorkKit"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "pyside6 (>=6.9.2,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pillow (==11.2.1)",
    "numpy (==1.24.4)",
    "attrs (>=25.3.0,<26.0.0)",
    "sqlalchemy (==2.0.40)",
    "alembic (==1.12.1)",
    "pynsq (>=0.9.1,<0.10.0)",
    "pyinstaller (==6.15.0)",
    "pymupdf (==1.22.5)",
    "pypdf (==4.3.1)",
    "reportlab (==4.2.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[tool.poetry]
package-mode = false
